fix: screen error on playerstyle change
sepcifcically ApperanceSettings cant have a portal (here its checkDialog)
